Extract execution planner into a submodule
Extract the execution planning methods out of `Graph` and into a separate
`Planner` type in a `planner` submodule. Also move the associated `CaptureEnv`
and `PlanOptions` types there.

This is another step in breaking down the large `graph` module into more
manageable pieces.
